Feat: refactor and add tests.
