+++ transforms/images/rolling-ball-plugin/tests/test_cli.py
@@ -2,65 +2,21 @@
 
 import faulthandler
 import json
-import tempfile
-from pathlib import Path
-
-import numpy
-import pytest
-from bfio import BioWriter
 from typer.testing import CliRunner
 
-from src.polus.plugins.transforms.images.rolling_ball.__main__ import app as app
+from polus.plugins.transforms.images.rolling_ball.__main__ import app as app
+
+from fixtures import (
+    paths,
+    image_data,
+    image_file
+)
+
 
 faulthandler.enable()
 
-
-def get_temp_file(path: Path, suffix: str):
-    """Create path to a temp file."""
-    temp_name = next(tempfile._get_candidate_names())
-    return path / (temp_name + suffix)
-
-
-@pytest.fixture()
-def paths(tmp_path):
-    """Create temporary directories"""
-    input_dir = tmp_path / "inp"
-    output_dir = tmp_path / "out"
-    input_dir.mkdir()
-    output_dir.mkdir()
-    return (input_dir, output_dir)
-
-
-@pytest.fixture()
-def image():
-    """Create a numpy array to generate an input image."""
-    image_size = 64
-    image_shape = (image_size, image_size)
-    random_image = numpy.random.randint(
-        low=0,
-        high=255,
-        size=image_shape,
-        dtype=numpy.uint8,
-    )
-    return random_image
-
-
-@pytest.fixture()
-def infile(image, paths):
-    """Create an image file in the input directory."""
-    inp_dir, _ = paths
-    infile = get_temp_file(inp_dir, "ome.tiff")
-
-    with BioWriter(infile) as writer:
-        writer.X = image.shape[0]
-        writer.Y = image.shape[1]
-        writer[:] = image[:]
-
-    return infile
-
-
-def test_cli(infile, paths):  # noqa
-    """Test the Rolling Ball plugin from the command line."""
+def test_cli(image_file, paths):  # noqa
+    """Test the command line exit code with valid inputs."""
     runner = CliRunner()
 
     inp_dir, out_dir = paths
@@ -73,7 +29,7 @@ def test_cli(infile, paths):  # noqa
     assert result.exit_code == 0
 
 
-def test_cli_preview(infile, paths, caplog):  # noqa
+def test_cli_preview(image_file, paths):  # noqa
     """Test the client with the preview option."""
 
     runner = CliRunner()
@@ -94,11 +50,11 @@ def test_cli_preview(infile, paths, caplog):  # noqa
     # verify we generate the preview file
     result = plugin_json["outputDir"]
     assert len(result) == 1
-    assert result[0] == infile.name
+    assert result[0] == image_file.name
 
 
 def test_cli_bad_input(paths):  # noqa
-    """Test the Rolling Ball plugin from the command line."""
+    """"Test bad inputs."""
 
     runner = CliRunner()
